+++ libsolidity/ast/AST.h
@@ -423,6 +423,9 @@ public:
 	/// Returns the fallback function or nullptr if no fallback function was specified.
 	FunctionDefinition const* fallbackFunction() const;
 
+	/// Returns the ether receiver function or nullptr if no ether receiver function was specified.
+	FunctionDefinition const* etherReceiverFunction() const;
+
 	std::string fullyQualifiedName() const { return sourceUnitName() + ":" + name(); }
 
 	TypePointer type() const override;
@@ -638,7 +641,7 @@ public:
 		ASTPointer<ASTString> const& _name,
 		Declaration::Visibility _visibility,
 		StateMutability _stateMutability,
-		bool _isConstructor,
+		Token _kind,
 		ASTPointer<OverrideSpecifier> const& _overrides,
 		ASTPointer<ASTString> const& _documentation,
 		ASTPointer<ParameterList> const& _parameters,
@@ -650,27 +653,31 @@ public:
 		Documented(_documentation),
 		ImplementationOptional(_body != nullptr),
 		m_stateMutability(_stateMutability),
-		m_isConstructor(_isConstructor),
+		m_kind(_kind),
 		m_overrides(_overrides),
 		m_functionModifiers(_modifiers),
 		m_body(_body)
-	{}
+	{
+		solAssert(_kind == Token::Constructor || _kind == Token::Function || _kind == Token::Fallback || _kind == Token::Receive, "");
+	}
 
 	void accept(ASTVisitor& _visitor) override;
 	void accept(ASTConstVisitor& _visitor) const override;
 
 	StateMutability stateMutability() const { return m_stateMutability; }
-	bool isConstructor() const { return m_isConstructor; }
+	bool isConstructor() const { return m_kind == Token::Constructor; }
 	ASTPointer<OverrideSpecifier> const& overrides() const { return m_overrides; }
-	bool isFallback() const { return !m_isConstructor && name().empty(); }
+	bool isFallback() const { return m_kind == Token::Fallback; }
+	bool isEtherReceiver() const { return m_kind == Token::Receive; }
+	Token kind() const { return m_kind; }
 	bool isPayable() const { return m_stateMutability == StateMutability::Payable; }
 	std::vector<ASTPointer<ModifierInvocation>> const& modifiers() const { return m_functionModifiers; }
 	Block const& body() const { solAssert(m_body, ""); return *m_body; }
 	bool isVisibleInContract() const override
 	{
-		return Declaration::isVisibleInContract() && !isConstructor() && !isFallback();
+		return Declaration::isVisibleInContract() && !isConstructor() && !isFallback() && !isEtherReceiver();
 	}
-	bool isPartOfExternalInterface() const override { return isPublic() && !isConstructor() && !isFallback(); }
+	bool isPartOfExternalInterface() const override { return isPublic() && !isConstructor() && !isFallback() && !isEtherReceiver(); }
 
 	/// @returns the external signature of the function
 	/// That consists of the name of the function followed by the types of the
@@ -689,7 +696,7 @@ public:
 
 private:
 	StateMutability m_stateMutability;
-	bool m_isConstructor;
+	Token m_kind;
 	ASTPointer<OverrideSpecifier> m_overrides;
 	std::vector<ASTPointer<ModifierInvocation>> m_functionModifiers;
 	ASTPointer<Block> m_body;