+++ libsolidity/codegen/ir/IRGenerator.cpp
@@ -267,6 +267,7 @@ string IRGenerator::dispatchRoutine(ContractDefinition const& _contract)
 			</cases>
 			default {}
 		}
+		if iszero(calldatasize()) { <receiveEther> }
 		<fallback>
 	)X");
 	t("shr224", m_utils.shiftRightFunction(224));
@@ -313,6 +314,11 @@ string IRGenerator::dispatchRoutine(ContractDefinition const& _contract)
 	}
 	else
 		t("fallback", "revert(0, 0)");
+	if (FunctionDefinition const* etherReceiver = _contract.etherReceiverFunction())
+		// TODO: should this revert for iszero(callvalue())?
+		t("receiveEther", generateFunction(*etherReceiver) + "() stop()");
+	else
+		t("receiveEther", "");
 	return t.render();
 }